{"schema":"decodex/commit/1","summary":"split docs okf commands and Swift panel views","authority":"manual"}
